refactor: autodoc: Move deprecated documenters to autodoc.deprecated package
